+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ General
 - Added codacy integration (#6)
 - Reworked github CI workflow (#6)
 - Added ``popper.js`` library for tooltip support (#64)
+- Added ldap3 dependency (#60)
 
 UserSec
 -------
@@ -30,11 +31,12 @@ UserSec
 - Created progressing user view (#6)
 - Created group create request view (#7)
 - Added ``editor`` field to request models (#37)
-- Created HPC user view (#13)
+- Created HPC user overview view (#13)
 - Created HPC group detail view (#53)
 - Registered models with admin site (#53)
 - Created HPC user create request views (#58)
 - Reworked overview page (#64)
+- Created HPC user detail view (#60)
 
 AdminSec
 --------
@@ -46,3 +48,8 @@ AdminSec
 - Accepting group create request creates HPC user and group (#13)
 - Created HPC user create request views (#58)
 - Reworked overview page (#64)
+- Created HPC user detail view (#60)
+- Created HPC group detail view (#60)
+- Added logic for approving HPC user create request (#60)
+- Added LDAP Connector class to fetch username by email alongside LDAP authentication (#60)
+- Enabled email sending when HPC user create request is accepted (#60)